+++ packages/web3-telemetry/src/entry-types.ts
@@ -1,0 +1,1 @@
+export * from './types/index.js'
