+++ packages/web3-telemetry/src/hooks/useMountReport.ts
@@ -1,15 +1,15 @@
 import { useMount } from 'react-use'
 import { type Unresolved, resolve } from '@masknet/shared-base'
-import { TelemetryAPI } from '@masknet/web3-providers/types'
 import { useTelemetry } from './useTelemetry.js'
+import { type EventID, EventType } from '../types/index.js'
 
 /**
  * Log an access event
  */
-export function useMountReport(eventID: Unresolved<TelemetryAPI.EventID>) {
+export function useMountReport(eventID: Unresolved<EventID>) {
     const telemetry = useTelemetry()
 
     useMount(() => {
-        telemetry.captureEvent(TelemetryAPI.EventType.Access, resolve(eventID))
+        telemetry.captureEvent(EventType.Access, resolve(eventID))
     })
 }
