+++ packages/web3-telemetry/src/hooks/useUnmountReport.ts
@@ -1,15 +1,15 @@
 import { useUnmount } from 'react-use'
-import { TelemetryAPI } from '@masknet/web3-providers/types'
 import { type Unresolved, resolve } from '@masknet/shared-base'
+import { type EventID, EventType } from '../types/index.js'
 import { useTelemetry } from './useTelemetry.js'
 
 /**
  * Log an exit event
  */
-export function useUnmountReport(eventID: Unresolved<TelemetryAPI.EventID>) {
+export function useUnmountReport(eventID: Unresolved<EventID>) {
     const telemetry = useTelemetry()
 
     useUnmount(() => {
-        telemetry.captureEvent(TelemetryAPI.EventType.Exit, resolve(eventID))
+        telemetry.captureEvent(EventType.Exit, resolve(eventID))
     })
 }
