+++ packages/web3-telemetry/tsconfig.json
@@ -6,10 +6,5 @@
     "tsBuildInfoFile": "./dist/.tsbuildinfo"
   },
   "include": ["./src", "./src/**/*.json"],
-  "references": [
-    { "path": "../web3-shared/base" },
-    { "path": "../web3-hooks/base" },
-    { "path": "../web3-helpers" },
-    { "path": "../web3-providers" }
-  ]
+  "references": [{ "path": "../web3-shared/base" }, { "path": "../web3-hooks/base" }, { "path": "../web3-helpers" }]
 }
